server: compile with -Wextra, fix all warnings
Disable missing field initializers warning which is seen when
initializing some libwebsockets stuff. It's using the C method of
zeroing out unspecified fields, but apparently that's a warning in C++.

Disable unused parameter warnings, as I'm not about to go changing
function signatures.

Besides that, add various typecasts to suppress signed vs unsigned
comparisons in libwebsockets code.

+++ server/libwebsockets/libwebsockets.c
@@ -894,7 +894,7 @@ libwebsocket_service_fd(struct libwebsocket_context *context,
 
 	gettimeofday(&tv, NULL);
 
-	if (context->last_timeout_check_s != tv.tv_sec) {
+	if (context->last_timeout_check_s != (unsigned long)tv.tv_sec) {
 		context->last_timeout_check_s = tv.tv_sec;
 
 		#ifndef WIN32
